refactor(policy): retire dead policy check modules from the #1500 teardown
