+++ tools/policy/authz_matrix.py
@@ -1,4 +1,4 @@
-"""Policy checks: authorization matrix sync and PR body contracts.
+"""Policy checks: PR body contracts.
 
 Extracted from tools/policy/checks.py (issue #1355), which had reached 5,679 lines against
 the repo's 500-LOC limit. checks.py remains the entry point and re-exports this module, so
@@ -30,154 +30,6 @@ from .core import (
     extract_requirement_uids_section,
     run_no_deferral_disposition_check,
 )
-
-
-def _check_enforcement_target(rel: str, entry_id: str, target: object, root: Path) -> list[Violation]:
-    if not isinstance(target, str):
-        return [Violation(code="contract-invariant-enforcement-invalid", message=f"{rel} invariant {entry_id} has an invalid enforcement path.", details=[str(target)])]
-    target_path_text, separator, target_anchor = target.partition("::")
-    target_path = Path(target_path_text)
-    if not separator or not target_anchor:
-        return [Violation(code="contract-invariant-enforcement-anchor-missing", message=f"{rel} invariant {entry_id} must name a specific test/spec anchor.", details=[f"use '<repo-path>::<test-or-rule-id>', got {target}"])]
-    if target_path_text.startswith("/") or ".." in target_path.parts:
-        return [Violation(code="contract-invariant-enforcement-invalid", message=f"{rel} invariant {entry_id} has an invalid enforcement path.", details=[target])]
-    resolved = root / target_path
-    if not resolved.exists():
-        return [Violation(code="contract-invariant-enforcement-missing-file", message=f"{rel} invariant {entry_id} references a missing enforcement file.", details=[target_path_text])]
-    if target_anchor not in resolved.read_text(encoding="utf-8"):
-        return [Violation(code="contract-invariant-enforcement-anchor-missing-file", message=f"{rel} invariant {entry_id} references an enforcement anchor that is not present.", details=[target])]
-    return []
-
-
-def _check_invariant_entry(rel: str, entry: object, root: Path) -> list[Violation]:
-    if not isinstance(entry, dict) or not entry.get("id"):
-        return [Violation(code="contract-invariant-entry-invalid", message=f"{rel} has an invariant entry without an id.")]
-    if entry["id"] == "none":
-        if not entry.get("rationale"):
-            return [Violation(code="contract-invariant-none-rationale-missing", message=f"{rel} declares no invariants but omits a rationale.")]
-        return []
-    enforced_by = entry.get("enforcedBy")
-    if not isinstance(enforced_by, list) or len(enforced_by) == 0:
-        return [Violation(code="contract-invariant-enforcement-missing", message=f"{rel} invariant {entry['id']} must name at least one enforcing test or spec file.")]
-    violations: list[Violation] = []
-    for target in enforced_by:
-        violations.extend(_check_enforcement_target(rel, entry["id"], target, root))
-    return violations
-
-
-def _check_schema_invariants(rel: str, schema: dict, root: Path) -> list[Violation]:
-    invariants = schema.get("x-ground-control-invariants")
-    if invariants is None:
-        return [Violation(code="contract-invariant-inventory-missing", message=f"{rel} must declare x-ground-control-invariants.", details=["Use [{\"id\":\"none\", \"rationale\":\"...\"}] only when the schema has no declared invariant."])]
-    if not isinstance(invariants, list) or len(invariants) == 0:
-        return [Violation(code="contract-invariant-inventory-invalid", message=f"{rel} has an empty or non-list x-ground-control-invariants value.")]
-    violations: list[Violation] = []
-    for entry in invariants:
-        violations.extend(_check_invariant_entry(rel, entry, root))
-    return violations
-
-
-def run_contract_invariant_enforcement_check(root: Path = REPO_ROOT) -> list[Violation]:
-    schemas_dir = root / "contracts" / "schemas"
-    if not schemas_dir.exists():
-        return [Violation(code="contract-schema-dir-missing", message="contracts/schemas/ is missing; GC-O014 schema invariant coverage cannot be checked.")]
-    violations: list[Violation] = []
-    for schema_path in sorted(schemas_dir.rglob("*.schema.json")):
-        rel = schema_path.relative_to(root).as_posix()
-        try:
-            schema = json.loads(schema_path.read_text(encoding="utf-8"))
-        except json.JSONDecodeError as exc:
-            violations.append(Violation(code="contract-schema-json-invalid", message=f"{rel} is not valid JSON.", details=[str(exc)]))
-            continue
-        violations.extend(_check_schema_invariants(rel, schema, root))
-    return violations
-
-def _parse_authz_contract_rows(text: str) -> list[dict[str, str]]:
-    rows: list[dict[str, str]] = []
-    current: dict[str, str] = {}
-    for raw_line in text.splitlines():
-        line = raw_line.strip()
-        if line.startswith("- id:"):
-            if current:
-                rows.append(current)
-            current = {"id": line.split(":", 1)[1].strip().strip('"')}
-        elif current and ":" in line:
-            key, value = line.split(":", 1)
-            current[key.strip()] = value.strip().strip('"')
-    if current:
-        rows.append(current)
-    return rows
-
-
-def _java_matrix_paths_by_access(java_text: str) -> dict[str, set[str]]:
-    constants = dict(re.findall(r'private\s+static\s+final\s+String\s+(\w+)\s*=\s*"([^"]+)"', java_text))
-    access_methods = {
-        "hasRole(ROLE_ADMIN)": "ROLE_ADMIN",
-        "access(identityAuthorizationManager)": "PERMISSION_IDENTITY_ADMIN",
-    }
-    paths_by_access = {access: set() for access in access_methods.values()}
-    matcher = r"\.requestMatchers\((.*?)\)\s*\.(hasRole\(ROLE_ADMIN\)|access\(identityAuthorizationManager\))"
-    for match in re.finditer(matcher, java_text, re.DOTALL):
-        block = match.group(1)
-        access = access_methods[match.group(2)]
-        paths = paths_by_access[access]
-        paths.update(re.findall(r'"(/api/v1/[^"]+)"', block))
-        for token in re.findall(r"\b[A-Z][A-Z0-9_]+\b", block):
-            if token in constants and constants[token].startswith("/api/v1/"):
-                paths.add(constants[token])
-    return paths_by_access
-
-
-def _java_admin_matrix_paths(java_text: str) -> set[str]:
-    return _java_matrix_paths_by_access(java_text)["ROLE_ADMIN"]
-
-
-def run_authz_matrix_sync_check(root: Path = REPO_ROOT) -> list[Violation]:
-    matrix_path = root / "contracts" / "authz" / "path-matrix.yaml"
-    java_path = root / "backend" / "src" / "main" / "java" / "com" / "keplerops" / "groundcontrol" / "shared" / "security" / "ApiPathMatrix.java"
-    if not matrix_path.exists() or not java_path.exists():
-        return [
-            Violation(
-                code="authz-matrix-source-missing",
-                message="Authz matrix sync check needs contracts/authz/path-matrix.yaml and ApiPathMatrix.java.",
-            )
-        ]
-
-    rows = _parse_authz_contract_rows(matrix_path.read_text(encoding="utf-8"))
-    contract_paths_by_access = {
-        access: {row.get("path", "") for row in rows if row.get("access") == access}
-        for access in ("ROLE_ADMIN", "PERMISSION_IDENTITY_ADMIN")
-    }
-    for paths in contract_paths_by_access.values():
-        paths.discard("")
-    java_paths_by_access = _java_matrix_paths_by_access(java_path.read_text(encoding="utf-8"))
-
-    violations: list[Violation] = []
-    details: list[str] = []
-    for access in ("ROLE_ADMIN", "PERMISSION_IDENTITY_ADMIN"):
-        contract_paths = contract_paths_by_access[access]
-        java_paths = java_paths_by_access[access]
-        missing_from_contract = sorted(java_paths - contract_paths)
-        missing_from_java = sorted(contract_paths - java_paths)
-        if missing_from_contract:
-            details.append(
-                f"{access} paths in ApiPathMatrix.java but not contracts/authz/path-matrix.yaml: "
-                f"{missing_from_contract}"
-            )
-        if missing_from_java:
-            details.append(
-                f"{access} paths in contracts/authz/path-matrix.yaml but not ApiPathMatrix.java: "
-                f"{missing_from_java}"
-            )
-    if details:
-        violations.append(
-            Violation(
-                code="authz-matrix-drift",
-                message="contracts/authz/path-matrix.yaml drifted from ApiPathMatrix.java.",
-                details=details,
-            )
-        )
-    return violations
 
 
 def run_pr_body_check(event_path: Path) -> list[Violation]:
