+++ tools/policy/checks.py
@@ -75,15 +75,8 @@ from .file_size import (  # noqa: F401
     run_file_size_limit_check,
 )
 from .adr_guard import (  # noqa: F401
-    JAVA_DOTTED_NAME_RE,
-    JAVA_IMPORT_RE,
-    JAVA_MAIN_SOURCE_PREFIX,
-    JAVA_TEST_SOURCE_PREFIX,
-    WEBMVCTEST_ANNOTATION_RE,
-    _CLASS_LITERAL_SUFFIX,
     _trigger_is_in_scope,
     changed_lines_for,
-    controller_fully_qualified_name,
     filter_matches,
     get_repo_relative_files,
     load_json,
@@ -91,38 +84,21 @@ from .adr_guard import (  # noqa: F401
     merge_base_or,
     read_changed_files,
     run_adr_guard,
-    run_controller_contracts,
     run_git,
-    test_covers_controller,
 )
-from .migration_policy import (  # noqa: F401
+from .version_mirror import (  # noqa: F401
     RELEASE_PLEASE_CONFIG,
     RELEASE_PLEASE_MANIFEST,
     RELEASE_PLEASE_ROOT_PACKAGE,
+    _DOCUMENTATION_COVERAGE_FIXTURE,
+    _DOCUMENTATION_SECTION_RE,
     _GENERIC_VERSION_ANNOTATION,
     _QUOTED_VERSION_RE,
-    _extract_compose_backend_env_entries,
     _extract_generic_version,
     _extract_json_version,
     _jsonpath_keys,
-    _migration_content_at_ref,
-    _resolve_baseline_ref,
-    git_diff_for_paths,
-    run_migration_policy,
-)
-from .version_mirror import (  # noqa: F401
-    _DOCUMENTATION_COVERAGE_FIXTURE,
-    _DOCUMENTATION_SECTION_RE,
     run_documentation_coverage_check,
     run_version_mirror_consistency_check,
-)
-from .ci_contract import (  # noqa: F401
-    CANONICAL_GHCR_NAMESPACE,
-    GHCR_IMAGE_REF_RE,
-    GHCR_NAMESPACE_INVENTORY,
-    run_ci_strictness_contract,
-    run_deploy_compose_credential_passthrough,
-    run_ghcr_namespace_drift,
 )
 from .requirement_specs import (  # noqa: F401
     FRONTMATTER_CODE,
@@ -177,116 +153,12 @@ from .repo_identity import (  # noqa: F401
     _yaml_placeholder_consumed_names,
     run_repo_identity_drift,
 )
-from .env_templates import (  # noqa: F401
-    _binds_within,
-    _run_env_template_consumer_check,
-    _spring_binding_index,
-    _spring_binds,
-)
-from .deploy_artifacts import (  # noqa: F401
-    BACKEND_METHODOLOGY_CATALOG_PATH,
-    FRONTEND_API_TYPES_PATH,
-    MCP_LIB_DIR,
-    MCP_LIB_PATH,
-    SKILL_METHODOLOGY_CATALOG_PATH,
-    _AUDIT_ENUM_STATE_DIR,
-    _BLOCK_COMMENT_RE,
-    _ENUM_CONSTANT_RE,
-    _ENUM_STATE_DIR,
-    _GRAPH_MODEL_DIR,
-    _JAVA_ENUM_BODY_RE,
-    _LINE_COMMENT_RE,
-    _METHODOLOGY_BACKEND_SOURCE_RE,
-    _METHODOLOGY_METHOD_KEY_RE,
-    _METHODOLOGY_SKILL_SOURCE_RE,
-    _PAREN_GROUP_RE,
-    _STRING_LITERAL_RE,
-    _VERIFICATION_ENUM_STATE_DIR,
-    _parse_methodology_catalog,
-    read_mcp_library,
-    run_deploy_artifact_consistency,
-    run_methodology_catalog_drift,
-)
-from .enum_contract import (  # noqa: F401
-    ENUM_CONTRACT_INVENTORY,
-    EnumContract,
-    ONTOLOGY_CONTRACT_PATHS,
-    ONTOLOGY_OWNERS,
-    ONTOLOGY_PROVENANCE,
-    ONTOLOGY_SCHEMA_VERSIONS,
-    ONTOLOGY_SOURCE_ROOT,
-    ONTOLOGY_SURFACE_KINDS,
-    ONTOLOGY_TERM_KINDS,
-    _drift_violation,
-    _load_ontology_contracts,
-    _nonempty_string_list,
-    _ontology_violation,
-    _strip_comments,
-    parse_const_string_array,
-    parse_java_enum_constants,
-    parse_ts_union_literals,
-    run_enum_contract_check,
-)
-from .ontology_families import (  # noqa: F401
-    _contributor_edge_values,
-    _contributor_type_identity,
-    _edge_enum_selector,
-    _graph_edge_argument_lists,
-    _is_enum_forwarded_parameter,
-    _is_ontology_enum,
-    _java_call_argument_lists,
-    _java_type_identity,
-    _split_java_arguments,
-    _validate_ontology_families,
-    _validate_ontology_terms,
-)
-from .ontology_binding import (  # noqa: F401
-    ONTOLOGY_CROSSWALK_PATH,
-    ONTOLOGY_CROSSWALK_SCHEMA_VERSION,
-    ONTOLOGY_CROSSWALK_TIME_FAMILY,
-    ONTOLOGY_EFFECT_VOCABULARY,
-    ONTOLOGY_EXTERNAL_SNAPSHOT_ROOT,
-    _discover_ontology_surfaces,
-    _load_ontology_family_ids,
-    _safe_ontology_external_path,
-    _safe_ontology_source_path,
-    run_ontology_binding_check,
-)
-from .ontology_crosswalk import (  # noqa: F401
-    EMITTER_SOURCE_PATHS,
-    IMPLEMENT_MODULE_DIR,
-    IMPLEMENT_MECHANICAL_PATH,
-    MEASUREMENT_RECORD_SCHEMA_PATH,
-    STATION_CATALOGUE_PATH,
-    _PHASE_MARKER_LITERAL_RE,
-    _STATION_BY_ACTION_ENTRY_RE,
-    _STATION_CALL_RE,
-    _STATION_ID_FIELD_RE,
-    _STATION_TABLE_RE,
-    _validate_crosswalk_pin,
-    run_contract_surface_check,
-    run_ontology_crosswalk_check,
-)
-from .measurement import (  # noqa: F401
-    _catalogue_index,
-    _check_emitter_station_drift,
-    _check_phase_marker_drift,
-    _check_routing_stage_drift,
-    _load_station_catalogue,
-    _resolves,
-    run_measurement_catalogue_check,
-)
 from .authz_matrix import (  # noqa: F401
     IMPLEMENT_SKILL_PATH,
     IMPLEMENT_STEP_TEST_QUALITY_PATH,
     _STEP_HEADING_RE,
-    _java_admin_matrix_paths,
-    _java_matrix_paths_by_access,
-    _parse_authz_contract_rows,
     check_pr_body,
     extract_step_section,
-    run_authz_matrix_sync_check,
-    run_contract_invariant_enforcement_check,
     run_pr_body_check,
 )
 from .decision_records import (  # noqa: F401
@@ -297,17 +169,20 @@ from .decision_records import (  # noqa: F401
 from .workflow_routing import (  # noqa: F401
     parse_args,
     render_and_exit,
-    run_traceability_reconciliation_gate_contract,
     run_workflow_routing_contract,
     write_violations_json,
 )
 from .execution_contract import (  # noqa: F401
+    MCP_LIB_DIR,
+    MCP_LIB_PATH,
     RELEASE_PLEASE_PR_HEAD_PREFIX,
     RELEASE_PR_BASE,
     RELEASE_PR_HEAD,
     SYNC_PR_BASE,
     SYNC_PR_HEAD,
     _resolve_pr_body,
+    read_mcp_library,
+    read_mcp_registrations,
     run_implement_execution_contract,
 )
 from .cli import (  # noqa: F401
