+++ tools/tests/policy_fixtures.py
@@ -18,48 +18,8 @@ from unittest import mock
 from unittest.mock import patch
 
 from tools.policy.checks import (
-    DEFERRAL_CASES_PATH,
-    ENUM_CONTRACT_INVENTORY,
-    FRONTEND_API_TYPES_PATH,
-    MCP_LIB_PATH,
     REPO_ROOT,
-    Violation,
-    _is_release_pr,
-    _jsonpath_keys,
-    _resolve_pr_refs,
-    check_pr_body,
-    classify_deferral_language,
-    extract_step_section,
     main,
-    parse_args,
-    parse_const_string_array,
-    parse_java_enum_constants,
-    parse_ts_union_literals,
-    read_changed_files,
-    run_adr_guard,
-    _trigger_is_in_scope,
-    run_ci_strictness_contract,
-    run_authz_matrix_sync_check,
-    run_contract_invariant_enforcement_check,
-    run_contract_surface_check,
-    run_controller_contracts,
-    run_deploy_artifact_consistency,
-    run_deploy_compose_credential_passthrough,
-    run_documentation_coverage_check,
-    run_enum_contract_check,
-    run_ghcr_namespace_drift,
-    run_repo_identity_drift,
-    run_measurement_catalogue_check,
-    run_migration_policy,
-    run_no_deferral_disposition_check,
-    run_ontology_binding_check,
-    run_ontology_crosswalk_check,
-    run_pr_body_check,
-    run_test_quality_decision_record_contract,
-    run_traceability_reconciliation_gate_contract,
-    run_version_mirror_consistency_check,
-    run_workflow_routing_contract,
-    run_implement_execution_contract,
 )
 
 
@@ -179,28 +139,6 @@ class PolicyChecksFixture(unittest.TestCase):
         return rel
 
     @staticmethod
-    def _init_migration_repo(tmp_dir, baseline_content):
-        """Create a git repo with one migration committed on a `baseline` ref."""
-        import subprocess
-
-        root = Path(tmp_dir)
-        rel = "backend/src/main/resources/db/migration/V100__baseline.sql"
-        path = root / rel
-        path.parent.mkdir(parents=True, exist_ok=True)
-        path.write_text(baseline_content, encoding="utf-8")
-
-        def git(*args):
-            subprocess.run(["git", *args], cwd=str(root), check=True, capture_output=True, text=True)
-
-        git("init", "-q")
-        git("config", "user.email", "t@example.com")
-        git("config", "user.name", "t")
-        git("add", "-A")
-        git("commit", "-q", "-m", "baseline")
-        git("branch", "baseline")
-        return root, rel, path
-
-    @staticmethod
     def _run_main_for_event(base_ref, head_ref, body):
         import contextlib
         import io
@@ -227,173 +165,3 @@ class PolicyChecksFixture(unittest.TestCase):
             "- [x] Pre-push code review and test-quality review completed; all findings fixed or dispositioned\n"
             "## Traceability\n\n- IMPLEMENTS: foo\n- TESTS: bar\n"
         )
-
-    def _write_valid_deploy_tree(self, root: Path) -> None:
-        """Write a minimal internally-consistent deploy artifact set.
-
-        Mirrors the invariants run_deploy_artifact_consistency enforces so a
-        test can introduce exactly one defect and assert the matching code.
-        """
-        ddir = root / "deploy/docker"
-        ddir.mkdir(parents=True, exist_ok=True)
-        (ddir / ".env.example").write_text(
-            "GC_IMAGE=ghcr.io/autarchy-ai/ground-control:main\n", encoding="utf-8"
-        )
-        (ddir / "env.schema").write_text(
-            "REQUIRED GC_IMAGE\n"
-            "RELEASE_PIN GC_IMAGE\n"
-            "REQUIRED GC_DATABASE_URL\n"
-            "REQUIRED GC_BIND_IP\n",
-            encoding="utf-8",
-        )
-        (ddir / "docker-compose.prod.yml").write_text(
-            "services:\n"
-            "  backend:\n"
-            "    image: ${GC_IMAGE}\n"
-            "    environment:\n"
-            "      - GC_DATABASE_URL=${GC_DATABASE_URL}\n"
-            "    ports:\n"
-            "      - \"${GC_BIND_IP}:8000:8000\"\n",
-            encoding="utf-8",
-        )
-        (ddir / "deploy.sh").write_text("#!/bin/bash\ndocker compose --env-file .env up -d\n", encoding="utf-8")
-        (ddir / "validate-env.sh").write_text("#!/bin/bash\nexit 0\n", encoding="utf-8")
-        # Manifest must match the four canonical artifacts byte-for-byte.
-        manifest_lines = []
-        for name in ("deploy.sh", "docker-compose.prod.yml", "validate-env.sh", "env.schema"):
-            digest = hashlib.sha256((ddir / name).read_bytes()).hexdigest()
-            manifest_lines.append(f"{digest}  {name}")
-        (ddir / "MANIFEST.sha256").write_text("\n".join(manifest_lines) + "\n", encoding="utf-8")
-        wrapper = root / "scripts/deploy.sh"
-        wrapper.parent.mkdir(parents=True, exist_ok=True)
-        wrapper.write_text("#!/bin/bash\nssh gc-deploy@red-dragon\n", encoding="utf-8")
-
-    def _rewrite_manifest(self, root: Path) -> None:
-        ddir = root / "deploy/docker"
-        lines = []
-        for name in ("deploy.sh", "docker-compose.prod.yml", "validate-env.sh", "env.schema"):
-            digest = hashlib.sha256((ddir / name).read_bytes()).hexdigest()
-            lines.append(f"{digest}  {name}")
-        (ddir / "MANIFEST.sha256").write_text("\n".join(lines) + "\n", encoding="utf-8")
-
-    def _write_spring_properties(self, root: Path) -> None:
-        # Mirrors the real SecurityProperties: the indexed credential slots bind
-        # through a nested POJO, so the fixture must carry that POJO's leaf
-        # fields — the binding path is resolved to the leaf, not to `credentials`.
-        props = root / "backend/src/main/java/com/example/SecurityProperties.java"
-        props.parent.mkdir(parents=True, exist_ok=True)
-        props.write_text(
-            '@ConfigurationProperties(prefix = "groundcontrol.security", '
-            "ignoreUnknownFields = false)\n"
-            "public class SecurityProperties {\n"
-            "    private boolean enabled = true;\n"
-            "    private List<ApiCredential> credentials = new ArrayList<>();\n"
-            "    private List<String> ipAllowlist = new ArrayList<>();\n"
-            "\n"
-            "    public static class ApiCredential {\n"
-            "        private String principalName;\n"
-            "        private String token;\n"
-            "        private Role role;\n"
-            "    }\n"
-            "}\n",
-            encoding="utf-8",
-        )
-
-    def _copy_enum_sources(self, root: Path) -> None:
-        rels = [FRONTEND_API_TYPES_PATH, MCP_LIB_PATH, *[c.java_path for c in ENUM_CONTRACT_INVENTORY]]
-        for rel in rels:
-            src = REPO_ROOT / rel
-            dst = root / rel
-            dst.parent.mkdir(parents=True, exist_ok=True)
-            dst.write_text(src.read_text(encoding="utf-8"), encoding="utf-8")
-
-    def _write_minimal_ontology_fixture(self, root: Path) -> None:
-        java_root = root / "backend" / "src" / "main" / "java" / "example"
-        java_root.mkdir(parents=True)
-        (java_root / "GraphEntityType.java").write_text(
-            "package example; public enum GraphEntityType { REQUIREMENT }\n",
-            encoding="utf-8",
-        )
-        (java_root / "RequirementGraphProjectionContributor.java").write_text(
-            """package example;
-public class RequirementGraphProjectionContributor implements GraphProjectionContributor {
-    Object edge() { return new GraphEdge("id", "RELATES", null, null, null, null, null); }
-}
-""",
-            encoding="utf-8",
-        )
-        ontology = root / "contracts" / "ontology"
-        ontology.mkdir(parents=True)
-        (ontology / "gc-concept-families-v1.json").write_text(
-            json.dumps(
-                {
-                    "schema_version": "gc-concept-families/v1",
-                    "owners": ["ground-control"],
-                    "families": {
-                        "requirements-and-traceability": {
-                            "title": "Requirements and traceability",
-                            "description": "Requirement identity and relationships.",
-                            "provenance": "native",
-                            "owner": "ground-control",
-                            "extension_scope": "Requirement and traceability concepts.",
-                            "relation_rules": ["Relations preserve declared direction."],
-                            "non_ambiguity_constraints": ["Requirements are not evidence artifacts."],
-                        }
-                    },
-                }
-            ),
-            encoding="utf-8",
-        )
-        (ontology / "gc-controlled-vocabularies-v1.json").write_text(
-            json.dumps(
-                {
-                    "schema_version": "gc-controlled-vocabularies/v1",
-                    "terms": {
-                        "node.requirement": {
-                            "kind": "classification",
-                            "title": "Requirement",
-                            "description": "A governed requirement.",
-                            "family": "requirements-and-traceability",
-                            "owner": "ground-control",
-                        },
-                        "edge.relates": {
-                            "kind": "edge",
-                            "title": "Relates",
-                            "description": "A directed relation.",
-                            "family": "requirements-and-traceability",
-                            "owner": "ground-control",
-                            "direction": "source-to-target",
-                            "source_roles": ["source"],
-                            "target_roles": ["target"],
-                        },
-                    },
-                }
-            ),
-            encoding="utf-8",
-        )
-        (ontology / "gc-artifact-bindings-v1.json").write_text(
-            json.dumps(
-                {
-                    "schema_version": "gc-artifact-bindings/v1",
-                    "surfaces": [
-                        {
-                            "id": "example.GraphEntityType",
-                            "kind": "java-enum",
-                            "path": "backend/src/main/java/example/GraphEntityType.java",
-                            "bindings": [{"local_value": "REQUIREMENT", "term": "node.requirement"}],
-                        },
-                        {
-                            "id": "example.RequirementGraphProjectionContributor",
-                            "kind": "graph-contributor",
-                            "path": "backend/src/main/java/example/RequirementGraphProjectionContributor.java",
-                            "bindings": [{"local_value": "RELATES", "term": "edge.relates"}],
-                        },
-                    ],
-                }
-            ),
-            encoding="utf-8",
-        )
-
-    def _read_ontology_fixture_json(self, root: Path, filename: str) -> tuple[Path, dict]:
-        path = root / "contracts" / "ontology" / filename
-        return path, json.loads(path.read_text(encoding="utf-8"))
