+++ tools/tests/test_policy_implement_execution.py
@@ -14,8 +14,6 @@ from unittest.mock import patch
 
 from tools.policy.checks import (
     DEFERRAL_CASES_PATH,
-    ENUM_CONTRACT_INVENTORY,
-    FRONTEND_API_TYPES_PATH,
     MCP_LIB_PATH,
     REPO_ROOT,
     Violation,
@@ -27,31 +25,14 @@ from tools.policy.checks import (
     extract_step_section,
     main,
     parse_args,
-    parse_const_string_array,
-    parse_java_enum_constants,
-    parse_ts_union_literals,
     read_changed_files,
     run_adr_guard,
     _trigger_is_in_scope,
-    run_ci_strictness_contract,
-    run_authz_matrix_sync_check,
-    run_contract_invariant_enforcement_check,
-    run_contract_surface_check,
-    run_controller_contracts,
-    run_deploy_artifact_consistency,
-    run_deploy_compose_credential_passthrough,
     run_documentation_coverage_check,
-    run_enum_contract_check,
-    run_ghcr_namespace_drift,
     run_repo_identity_drift,
-    run_measurement_catalogue_check,
-    run_migration_policy,
     run_no_deferral_disposition_check,
-    run_ontology_binding_check,
-    run_ontology_crosswalk_check,
     run_pr_body_check,
     run_test_quality_decision_record_contract,
-    run_traceability_reconciliation_gate_contract,
     run_version_mirror_consistency_check,
     run_workflow_routing_contract,
     run_implement_execution_contract,
